+++ src/controllers/paid/traffic.js
@@ -15,6 +15,7 @@ import {
   notFound,
   forbidden,
   badRequest,
+  found,
 } from '@adobe/spacecat-shared-http-utils';
 import { AWSAthenaClient } from '@adobe/spacecat-shared-athena-client';
 import crypto from 'crypto';
@@ -23,31 +24,52 @@ import AccessControlUtil from '../../support/access-control-util.js';
 import { TrafficDataResponseDto } from '../../dto/traffic-data-base-response.js';
 import { TrafficDataWithCWVDto } from '../../dto/traffic-data-response-with-cwv.js';
 import {
-  parseCsvToJson,
   getS3CachedResult,
-  copyOneNewestCsvToCache,
+  addResultJsonToCache,
+  fileExists,
 } from './caching-helper.js';
+import { getDateRanges } from '../../../test/controllers/paid/calendar-week-helper.js';
 
 async function loadSql(variables) {
   return getStaticContent(variables, './src/controllers/paid/channel-query.sql.tpl');
 }
 
-async function tryGetCacheResult(siteid, s3, cacheBucket, query, log) {
-  const outPrefix = `${crypto.createHash('md5').update(query).digest('hex')}`;
-  const cacheKey = `${cacheBucket}/${siteid}/${outPrefix}.csv`;
-  const cached = await getS3CachedResult(s3, cacheKey, log);
-  if (cached) {
-    log.info(`Found cached result. Returning cached Athena result from S3: ${cacheKey}`);
-    const parsedJson = await parseCsvToJson(cached);
+function getWeekMonthsAndYears(year, week) {
+  const ranges = getDateRanges(week, year);
+  const months = [...new Set(ranges.map((r) => r.month))].join(', ');
+  const years = [...new Set(ranges.map((r) => r.year))].join(', ');
+  return { months, years };
+}
 
-    return { parsedJson, cacheKey, outPrefix };
-  }
-  return { parsedJson: null, cacheKey, outPrefix };
+function getCacheKey(siteId, query, cacheLocation) {
+  const outPrefix = crypto.createHash('md5').update(query).digest('hex');
+  const cacheKey = `${cacheLocation}/${siteId}/${outPrefix}.json`;
+  return { cacheKey, outPrefix };
 }
 
 function TrafficController(context, log, env) {
-  const { dataAccess } = context;
+  const { dataAccess, s3 } = context;
   const { Site } = dataAccess;
+
+  const {
+    RUM_METRICS_DATABASE: rumMetricsDatabase,
+    RUM_METRICS_COMPACT_TABLE: rumMetricsCompactTable,
+    S3_BUCKET_NAME: bucketName,
+  } = env;
+
+  // constants
+  const ATHENA_TEMP_FOLDER = `s3://${bucketName}/rum-metrics-compact/temp/out`;
+  const CACHE_LOCATION = `s3://${bucketName}/rum-metrics-compact/cache`;
+
+  async function tryGetCacheResult(siteId, query) {
+    const { cacheKey, outPrefix } = getCacheKey(siteId, query, CACHE_LOCATION);
+    if (await fileExists(s3, cacheKey, log)) {
+      log.info(`Found cached result. Fetching signed URL for Athena result from S3: ${cacheKey}`);
+      const cachedUrl = await getS3CachedResult(s3, cacheKey, log);
+      return { cachedResultUrl: cachedUrl, cacheKey, outPrefix };
+    }
+    return { cachedResultUrl: null, cacheKey, outPrefix };
+  }
 
   async function fetchPaidTrafficData(dimensions, mapper) {
     const siteId = context.params?.siteId;
@@ -60,87 +82,74 @@ function TrafficController(context, log, env) {
       return forbidden('Only users belonging to the organization can view paid traffic metrics');
     }
 
-    const {
-      siteKey, year, week, month,
-    } = context.data;
-    if (!siteKey || !year || !week || !month) {
-      const badReqMessagge = 'siteKey, year, month and week are required parameters';
-      log.info(badReqMessagge);
-      return badRequest(badReqMessagge);
+    // validate input params
+    const { year, week } = context.data;
+    if (!year || !week) {
+      return badRequest('Year and week are required parameters');
     }
-    const dbName = env.PAID_TRAFFIC_DATABASE;
-    const tableName = env.PAID_TRAFFIC_TABLE_NAME;
-    const fullTableName = `${dbName}.${tableName}`;
+
+    const { months, years } = getWeekMonthsAndYears(year, week);
+    const tableName = `${rumMetricsDatabase}.${rumMetricsCompactTable}`;
     const groupBy = dimensions;
     const dimensionColumns = groupBy.join(', ');
     const dimensionColumnsPrefixed = `${groupBy.map((col) => `a.${col}`).join(', ')}, `;
-    const description = `fetch paid channel data | db: ${dbName} | siteKey: ${siteKey} | year: ${year} | month: ${month} | week: ${week} | groupBy: [${groupBy.join(', ')}] | template: channel-query.sql.tpl`;
+    const description = `fetch paid channel data | db: ${rumMetricsDatabase} | siteKey: ${siteId} | year: ${year} | month: ${months} | week: ${week} | groupBy: [${groupBy.join(', ')}] | template: channel-query.sql.tpl`;
+
     log.info(`Processing query: ${description}`);
 
+    // build query
     const query = await loadSql({
-      siteId: siteKey,
-      year,
-      month,
+      siteId,
+      years,
+      months,
       week,
       groupBy: groupBy.join(', '),
       dimensionColumns,
       dimensionColumnsPrefixed,
-      tableName: fullTableName,
+      tableName,
     });
 
-    log.debug(`Fetching paid data with query ${query}`);
-    const outputFolder = env.PAID_TRAFFIC_S3_OUTPUT_URI || 's3://spacecat-dev-segments/temp/out';
-    const cacheBucket = env.PAID_TRAFFIC_S3_CACHE_BUCKET_URI || 's3://spacecat-dev-segments/cache';
-    const s3 = context.s3?.s3Client;
-    const {
-      parsedJson,
-      cacheKey,
-      outPrefix,
-    } = await tryGetCacheResult(siteId, s3, cacheBucket, query, log);
+    log.debug(`Fetching paid data with query: ${query}`);
+
+    // first try to get from cache
+    const { cachedResultUrl, cacheKey, outPrefix } = await tryGetCacheResult(siteId, query);
     const thresholdConfig = env.CWV_THRESHOLDS || {};
-    if (parsedJson) {
-      const responseCashed = parsedJson.map((row) => mapper.toJSON(row, thresholdConfig));
-      log.info(`Succesfully fetched results of size ${responseCashed?.length} from cache`);
-      return ok(responseCashed);
+    if (cachedResultUrl) {
+      log.info(`Successfully fetched presigned URL for cached result file: ${cacheKey}`);
+      return found(cachedResultUrl);
     }
 
-    const resultLocation = `${outputFolder}/${outPrefix}`;
+    // if not cached, query Athena
+    const resultLocation = `${ATHENA_TEMP_FOLDER}/${outPrefix}`;
     const athenaClient = AWSAthenaClient.fromContext(context, resultLocation);
 
-    log.info(`Fetching paid data directly from athena bucket ${outputFolder} and table ${fullTableName}`);
-    const results = await athenaClient.query(
-      query,
-      dbName,
-      description,
-    );
-
-    if (results && results.length > 0) {
-      const isCached = await copyOneNewestCsvToCache(s3, resultLocation, cacheKey, log);
-      log.info(`Is Copy Athena result CSV to S3 cche: ${cacheKey} succesful was : ${isCached}`);
-    }
+    log.info(`Fetching paid data directly from Athena table: ${tableName}`);
+    const results = await athenaClient.query(query, rumMetricsDatabase, description);
     const response = results.map((row) => mapper.toJSON(row, thresholdConfig));
-    log.info(`Succesfully fetched results of size ${response?.length}`);
-    return ok(response);
+    log.info(`Successfully fetched results of size ${response?.length}`);
+
+    // add to cache
+    let isCached = false;
+    if (response) {
+      isCached = await addResultJsonToCache(s3, cacheKey, response, log);
+      log.info(`Athena result JSON to S3 cache (${cacheKey}) successful: ${isCached}`);
+    }
+
+    log.warn(`Failed to cache result to S3 with key ${CACHE_LOCATION}. Returning response directly.`);
+    return ok(response, {
+      'content-encoding': 'gzip',
+    });
   }
 
-  const getPaidTrafficByCampaignUrlDevice = async () => fetchPaidTrafficData(['campaign', 'url', 'device'], TrafficDataWithCWVDto);
-  const getPaidTrafficByCampaignDevice = async () => fetchPaidTrafficData(['campaign', 'device'], TrafficDataWithCWVDto);
-  const getPaidTrafficByCampaignUrl = async () => fetchPaidTrafficData(['campaign', 'url'], TrafficDataWithCWVDto);
-  const getPaidTrafficByCampaign = async () => fetchPaidTrafficData(['campaign'], TrafficDataWithCWVDto);
-  const getPaidTrafficByTypeChannelCampaign = async () => fetchPaidTrafficData(['type', 'channel', 'campaign'], TrafficDataResponseDto);
-  const getPaidTrafficByTypeChannel = async () => fetchPaidTrafficData(['type', 'channel'], TrafficDataResponseDto);
-  const getPaidTrafficByTypeCampaign = async () => fetchPaidTrafficData(['type', 'campaign'], TrafficDataResponseDto);
-  const getPaidTrafficByType = async () => fetchPaidTrafficData(['type'], TrafficDataResponseDto);
-
   return {
-    getPaidTrafficByCampaignUrlDevice,
-    getPaidTrafficByCampaignDevice,
-    getPaidTrafficByCampaignUrl,
-    getPaidTrafficByCampaign,
-    getPaidTrafficByTypeChannelCampaign,
-    getPaidTrafficByTypeChannel,
-    getPaidTrafficByTypeCampaign,
-    getPaidTrafficByType,
+    getPaidTrafficByCampaignUrlDevice: () => fetchPaidTrafficData(['utm_campaign', 'path', 'device'], TrafficDataWithCWVDto),
+    getPaidTrafficByCampaignDevice: () => fetchPaidTrafficData(['utm_campaign', 'device'], TrafficDataWithCWVDto),
+    getPaidTrafficByCampaignUrl: () => fetchPaidTrafficData(['utm_campaign', 'path'], TrafficDataWithCWVDto),
+    getPaidTrafficByCampaign: () => fetchPaidTrafficData(['utm_campaign'], TrafficDataWithCWVDto),
+    getPaidTrafficByTypeChannelCampaign: () => fetchPaidTrafficData(['trf_type', 'trf_channel', 'utm_campaign'], TrafficDataResponseDto),
+    getPaidTrafficByTypeChannel: () => fetchPaidTrafficData(['trf_type', 'trf_channel'], TrafficDataResponseDto),
+    getPaidTrafficByTypeCampaign: () => fetchPaidTrafficData(['trf_type', 'utm_campaign'], TrafficDataResponseDto),
+    getPaidTrafficByType: () => fetchPaidTrafficData(['trf_type'], TrafficDataResponseDto),
   };
 }