+++ CHANGELOG.md
@@ -1,3 +1,66 @@
+# [1.135.0](https://github.com/adobe/spacecat-api-service/compare/v1.134.6...v1.135.0) (2025-07-21)
+
+
+### Features
+
+* introduce isSandbox flag for sandbox sites ([#1072](https://github.com/adobe/spacecat-api-service/issues/1072)) ([e4dca72](https://github.com/adobe/spacecat-api-service/commit/e4dca724328ecd968ead42352f2fad2dc6b12abb))
+
+## [1.134.6](https://github.com/adobe/spacecat-api-service/compare/v1.134.5...v1.134.6) (2025-07-20)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#1077](https://github.com/adobe/spacecat-api-service/issues/1077)) ([5922cf8](https://github.com/adobe/spacecat-api-service/commit/5922cf8e262d0fbddb83792139bdbbf95e030c01))
+
+## [1.134.5](https://github.com/adobe/spacecat-api-service/compare/v1.134.4...v1.134.5) (2025-07-19)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1076](https://github.com/adobe/spacecat-api-service/issues/1076)) ([aacb45b](https://github.com/adobe/spacecat-api-service/commit/aacb45b27e2918394c9c77d7f7ad3021dbc3c732))
+
+## [1.134.4](https://github.com/adobe/spacecat-api-service/compare/v1.134.3...v1.134.4) (2025-07-18)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-athena-client to v1.1.0 ([#1073](https://github.com/adobe/spacecat-api-service/issues/1073)) ([74ecf24](https://github.com/adobe/spacecat-api-service/commit/74ecf24b8fa9b7edb8a9e86f529bf7705e9ef6c5))
+
+## [1.134.3](https://github.com/adobe/spacecat-api-service/compare/v1.134.2...v1.134.3) (2025-07-18)
+
+
+### Bug Fixes
+
+* remove env vars and defaults ([#1074](https://github.com/adobe/spacecat-api-service/issues/1074)) ([6f5fcab](https://github.com/adobe/spacecat-api-service/commit/6f5fcab9c5e8d648d60178fd287f04103a7fd3e6))
+
+## [1.134.2](https://github.com/adobe/spacecat-api-service/compare/v1.134.1...v1.134.2) (2025-07-17)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.36.0 ([#1071](https://github.com/adobe/spacecat-api-service/issues/1071)) ([ee0b7ca](https://github.com/adobe/spacecat-api-service/commit/ee0b7cada195ac332157daded7efd4c55515e1e7))
+
+## [1.134.1](https://github.com/adobe/spacecat-api-service/compare/v1.134.0...v1.134.1) (2025-07-17)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#1068](https://github.com/adobe/spacecat-api-service/issues/1068)) ([208fa86](https://github.com/adobe/spacecat-api-service/commit/208fa86b0db6f0283ccdacfd66c19e6aec96559c))
+
+# [1.134.0](https://github.com/adobe/spacecat-api-service/compare/v1.133.2...v1.134.0) (2025-07-17)
+
+
+### Features
+
+* traffic-analyzer return link to file with redirect  ([#1067](https://github.com/adobe/spacecat-api-service/issues/1067)) ([7eda727](https://github.com/adobe/spacecat-api-service/commit/7eda727de641de4b4dafa6b0971317777c46538e))
+
+## [1.133.2](https://github.com/adobe/spacecat-api-service/compare/v1.133.1...v1.133.2) (2025-07-16)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#1065](https://github.com/adobe/spacecat-api-service/issues/1065)) ([9936d15](https://github.com/adobe/spacecat-api-service/commit/9936d1521c5b13f0b4b5a04a43034fb3f3fcc465))
+
 ## [1.133.1](https://github.com/adobe/spacecat-api-service/compare/v1.133.0...v1.133.1) (2025-07-15)